refactor(email): move llm_summary into yas.email, rename digest tests
Completes the yas.email consolidation:

- Move src/yas/alerts/digest/llm_summary.py -> src/yas/email/llm_summary.py
  (it's the digest top-line generator, the last real module under the old
  digest package). Update the two production importers and the test.
- Remove the now-empty src/yas/alerts/digest/ package entirely.
- Rename three test files to match where their code now lives:
    test_alerts_digest_builder.py     -> test_email_digest_builder.py
    test_alerts_digest_filters.py     -> test_email_filters.py
    test_alerts_digest_llm_summary.py -> test_email_llm_summary.py
- Refresh docstrings (filters/llm_summary are shared across all email
  templates now, not digest-specific) and add llm_summary.py to the
  yas/email/README.md directory map.

No behavior change; pure relocation + naming. 738 passed.
