--- src/yas/alerts/digest/__init__.py
@@ -1,1 +1,0 @@
-"""Digest builder package."""
